requirements.txt
aiofiles==22.1.0
asyncio==3.4.3
asyncpg==0.26.0
black==22.10.0
cffi==1.15.1
click==8.1.3
environs==9.5.0
greenlet==1.1.3.post0
httptools==0.5.0
install==1.3.5
marshmallow==3.18.0
multidict==6.0.2
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.1
platformdirs==2.5.2
pycparser==2.21
pycryptodome==3.15.0
PyJWT==2.6.0
pyparsing==3.0.9
python-dotenv==0.21.0
PyYAML==6.0
sanic==22.9.0
sanic-openapi==21.12.0
sanic-routing==22.8.0
secure==0.3.0
SQLAlchemy==1.4.42
tomli==2.0.1
ujson==5.5.0
uvloop==0.17.0
websockets==10.3
